+++ CHANGELOG.md
@@ -4,4 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Customize all stylelint rules
+- Update documentation
+
 [unreleased]: http://github.com/subchannel/css